+++ src/ledger/ReviewableRequestHelper.cpp
@@ -59,7 +59,7 @@ namespace stellar {
     }
 
     bool ReviewableRequestHelper::exists(Database &db, LedgerKey const &key) {
-        if (cachedEntryExists(key, db)) {
+        if (cachedEntryExists(key, db) && getCachedEntry(key, db)) {
             return true;
         }